pyproject.toml
[project]
name = "hermes-transaction-layer"
version = "0.1.0"
description = "Minimal transaction API service for a Hermes capability layer."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic-settings>=2.6.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["server/tests"]
pythonpath = ["."]

